Reconfiguration of all entities and DTO to adjust the modeling to improve the security of the application

+++ src/main/java/br/com/devannis/webmarket/model/dto/OrderItemsResponseDTO.java
@@ -1,9 +1,11 @@
 package br.com.devannis.webmarket.model.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemsResponseDTO(
         Long productId,
         String productName,
-        double productPrice,
+        BigDecimal productPrice,
         int quantity
 ) {
 }
